+++ packages/studio/src/components/editor/manualEdits.ts
@@ -20,6 +20,7 @@ export {
   readStudioPathOffset,
   readStudioBoxSize,
   readStudioRotation,
+  readGsapTranslateFromTransform,
   applyStudioPathOffset,
   applyStudioPathOffsetDraft,
   applyStudioBoxSize,
